+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [1.7.1-SNAPSHOT]
 
+- SortedFilteredList is editable and supports all functions of the ObservableList interface
 - ObservableXXXValue.onChange functions should support nullable values
 - Changed semantics of `Node.removeWhen` to switch visible/managed state instead of adding/removing from parent
 - Internal: ViewModel maintains a map between bound properties towards the ViewModel to support validators in a cleaner way without reflection calls to private APIs (https://github.com/edvin/tornadofx/issues/276)